Bug fixed: 日付の計算で桁あふれが発生する問題を解消。

+++ Implem.Pleasanter/Libraries/DataTypes/ProgressRate.cs
@@ -121,14 +121,14 @@ namespace Implem.Pleasanter.Libraries.DataTypes
                 : CreatedTime;
         }
 
-        private float PlannedValue(DateTime now, DateTime start, int range)
+        private float PlannedValue(DateTime now, DateTime start, double range)
         {
             return start < now && range > 0
                 ? (float)(Elapsed(now, start)) / (float)range
                 : 0;
         }
 
-        private static int Elapsed(DateTime now, DateTime start)
+        private static double Elapsed(DateTime now, DateTime start)
         {
             return Times.DateDiff(Times.Types.Seconds, start, now);
         }
